+++ docs/src/release_notes.rst
@@ -7,6 +7,8 @@ Release Notes
 =====
 + Updated ReadArgs decorator to return None when arg doesn't exist.
 + Updated ReadArgs to not log input value.
++ Updated testing framework to support permutations for Service Apps.
++ Added ``is_variable()`` method to Playbook module.
 
 2.0.2
 =====